% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_clip.R
\name{clip_ftw}
\alias{clip_ftw}
\title{Clip Fietstelweek data to a study area of interest}
\usage{
clip_ftw(directory, year, study_area, municipality = TRUE, ...,
  edges_file_2015 = NULL, nodes_file_2015 = NULL,
  routes_file_2015 = NULL, edges_file_2016 = NULL,
  nodes_file_2016 = NULL, routes_file_2016 = NULL)
}
\arguments{
\item{directory}{filepath to the directory where the Fietstelweek data is and will be stored.}

\item{year}{desired year of analysis, can either be \code{2015}, \code{2016}, or both \code{c(2015,2016)}.}

\item{study_area}{name of the municipality or province to clip the data to. The parameter is case sensitive
and prone to fail with spelling mistakes. To have a full list of the municipalities and
provinces names allowed, please type on the console \code{View(municipalities)} or
\code{View(provinces)}.}

\item{municipality}{logical value. Indicates if the study area is a municipality (\code{TRUE}) which is the
default, or a province (\code{FALSE}).}

\item{...}{ignored}

\item{edges_file_2015}{self downloaded filepath for the edges file for 2015. Its format should be compatible
with \link[sf]{st_read}.}

\item{nodes_file_2015}{self downloaded filepath for the nodes file for 2015. Its format should be compatible
with \link[sf]{st_read}.}

\item{routes_file_2015}{self downloaded filepath for the routes file for 2015. Its format should be compatible
with \link[data.table]{fread}.}

\item{edges_file_2016}{self downloaded filepath for the edges file for 2016. Its format should be compatible
with \link[sf]{st_read}.}

\item{nodes_file_2016}{self downloaded filepath for the nodes file for 2016. Its format should be compatible
with \link[sf]{st_read}.}

\item{routes_file_2016}{self downloaded filepath for the routes file for 2016. Its format should be compatible
with \link[data.table]{fread}.}
}
\value{
Prints the directory where the data has been saved.
}
\description{
Clips the Fietstelweek data on a system directory to a study area within the
Netherlands, either a municipality or a province for 2015 and/or 2016 and saves it as
three files: edges.gpkg, nodes.gpkg, routes.csv.
It can be used after applying the \link[ftw]{downloader_ftw} function, or providing
the user's self-downloaded data.
}
