% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_clip.R
\name{download_ftw}
\alias{download_ftw}
\title{Download function}
\usage{
download_ftw(directory, year, rm.zip)
}
\arguments{
\item{directory}{filepath to the directory where the Fietstelweek data will be stored.}

\item{year}{desired year of analysis, can either be \code{2015}, \code{2016}, or both \code{c(2015,2016)}.}

\item{rm.zip}{logical value to indicate if the zip files downloaded should be erased from the system (\code{TRUE})
which is the default, or kept (\code{FALSE})}
}
\description{
Download function for internal use.
}
