% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_clip.R
\name{downloader_clip_ftw}
\alias{downloader_clip_ftw}
\title{Download and clip Fietstelweek data to a study area of interest}
\usage{
downloader_clip_ftw(directory, year, study_area, municipality = TRUE,
  rm.src = TRUE)
}
\arguments{
\item{directory}{filepath to the directory where the clipped Fietstelweek data will be stored.}

\item{year}{desired year of analysis, can either be \code{2015}, \code{2016}, or both \code{c(2015,2016)}.}

\item{study_area}{name of the municipality or province to clip the data to. The parameter is case sensitive
and prone to fail with spelling mistakes. To have a full list of the municipalities and
provinces names allowed, please type on the console \code{View(municipalities)} or
\code{View(provinces)}.}

\item{municipality}{logical value. Indicates if the study area is a municipality (\code{TRUE}) which is the
default, or a province (\code{FALSE}).}

\item{rm.src}{logical value. Indicates if the complete Fietstelweek data should be removed from the system
(\code{TRUE}) which is the default, or kept (\code{FALSE}).}
}
\value{
Prints the directory where the data has been saved.
}
\description{
The function is a combination of \link[ftw]{downloader_ftw} and \link[ftw]{clip_ftw} functions.
It is specially made for the users who are only interested on a particular study area, and do not intend to use
the rest of the Dutch territory on their analysis.
}
\note{
This function might take a while to run. It depends on the Bikeprint server availability, and may sometimes fail to download the data.
      The function will return an error if this is the case and will ask the user to download the data manually.
}
