% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_clip.R
\name{downloader_ftw}
\alias{downloader_ftw}
\title{Download the complete Fietstelweek data}
\usage{
downloader_ftw(directory, year, rm.zip = TRUE)
}
\arguments{
\item{directory}{filepath to the directory where the Fietstelweek data will be stored.}

\item{year}{desired year of analysis, can either be \code{2015}, \code{2016}, or both \code{c(2015,2016)}.}

\item{rm.zip}{logical value to indicate if the zip files downloaded should be erased from the system (\code{TRUE})
which is the default, or kept (\code{FALSE})}
}
\value{
Prints the directory where the data has been saved.
}
\description{
Downloads the complete Fietstelweek data to a system directory for 2015 and/or 2016.
}
\note{
The function depends on the Bikeprint server availability, and may sometimes fail to download the data.
      The function will return an error if this is the case and will ask the user to download the data manually.
}
