% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\alias{explore.ftw_atmp}
\alias{explore.ftw_tmp}
\title{Explore the clipped Fietstelweek data}
\usage{
explore(obj, type, ..., network_coledges, network_colnodes, edges_coledges,
  nodes_colnodes, trips_coledges, trips_weekday, trips_hour, palette)

\method{explore}{ftw_atmp}(obj, type = "network", ...,
  network_coledges = "#ff8d00", network_colnodes = "#800080",
  edges_coledges = "intensity", nodes_colnodes = "waiting time",
  trips_coledges = "intensity", trips_weekday = c(0, 1, 2, 3, 4),
  trips_hour = NULL, palette = "plasma")

\method{explore}{ftw_tmp}(obj, type = "network", ...,
  network_coledges = "#ff8d00", network_colnodes = "#800080",
  edges_coledges = "intensity", nodes_colnodes = "waiting time",
  trips_coledges = "intensity", trips_weekday = c(0, 1, 2, 3, 4),
  trips_hour = NULL, palette = "plasma")
}
\arguments{
\item{obj}{an object of class \code{ftw_atmp} or \code{ftw_tmp}}

\item{type}{the basic information within the data that the user wants to explore.
Can be set to \code{network} (default), \code{edges}, \code{nodes}, \code{trips}.}

\item{...}{ignored}

\item{network_coledges}{if type is \code{network}, the parameter sets the color of the edges.}

\item{network_colnodes}{if type is \code{network}, the parameter sets the color of the nodes.}

\item{edges_coledges}{if type is \code{edges}, the parameter sets the color of the edges based on \code{intensity} (default) or \code{speed}.}

\item{nodes_colnodes}{if type is \code{nodes}, the parameter sets the color of the nodes based on \code{waiting time} (default) or \code{count}.}

\item{trips_coledges}{if type is \code{trips}, the parameter sets the color of the edges based on \code{intensity} (default) or \code{speed}.}

\item{trips_weekday}{if type is \code{trips}, the parameter subsets the data according to the day of the week. Monday is equivalent to 0 and Sunday to 6. The default shows weekdays \code{c(0,1,2,3,4)}.}

\item{trips_hour}{if type is \code{trips}, the parameter subsets the data according to the hour of the day. The default shows the whole day data.}

\item{palette}{sets the palette to color edges and nodes for types \code{edges}, \code{nodes}, and \code{trips}.}
}
\description{
Explore the clipped Fietstelweek data using the library tmap with its \code{view} mode.
}
