% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{ftw_sample_data}
\alias{ftw_sample_data}
\title{Fietstelweek data examples for Texel and Smallingerland, The Netherlands}
\format{The edges file is a geopackage with \code{MULTILINESTRING} geometry.
The nodes file is a geopackage with \code{POINT} geometry.
The routes file is a .csv file.
The attribute details for the three files can be found here:
\url{http://www.bikeprint.nl/fietstelweek/} under the name
"Bekijk documentatie bij de databastanden". The documentation is in Dutch.}
\source{
\url{http://www.bikeprint.nl/fietstelweek/}
}
\description{
Directory containing data for 2015 and 2016 downloaded and clipped with the
\code{ftw} package. The municipalities selected is Oss. The directory contains
an edge.gpkg, a node.gpkg, and a routes.csv file for each year.
}
\note{
The attributes for 2015 and 2016 are not completely equal.
}
