% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipalities}
\alias{municipalities}
\title{Municipalities names in the Netherlands}
\format{A matrix with 5 columns and 76 rows containing the names of the
        Dutch municipalities in alphabetical order.}
\source{
\url{http://www.imergis.nl/asp/47.asp}
}
\usage{
municipalities
}
\description{
A matrix containing the municipalities of
the Netherlands sorted in alfabetical order for referencial
use when clipping the Fietstelweek dataset.
}
\keyword{datasets}
