% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ftw}
\alias{plot_ftw}
\alias{plot.ftw_atmp}
\alias{plot_ftw}
\alias{plot.ftw_tmp}
\title{Plot the clipped Fietstelweek data}
\usage{
\method{plot}{ftw_atmp}(x, ..., type = "network",
  network_coledges = "grey50", network_colnodes = "cyan",
  trips_coledges = "intensity", trips_ncol = 3, palette = "plasma")

\method{plot}{ftw_tmp}(x, ..., network_coledges = "grey50",
  network_colnodes = "orange")
}
\arguments{
\item{x}{object of class \code{ftw_atmp} or \code{ftw_tmp}}

\item{...}{ignores}

\item{type}{the method allows for two type of plots. A basic one which plots the
\code{network} (default) and a facet plot of type \code{trips per day} that allows
to plot the edges according to intensity or speed for the different days of the week.}

\item{network_coledges}{if type is \code{network}, the parameter sets the color of the edges.}

\item{network_colnodes}{if type is \code{network}, the parameter sets the color of the nodes.}

\item{trips_coledges}{if type \code{trips per day}, the parameter sets the color of the edges based on \code{intensity} (default) or \code{speed}.}

\item{trips_ncol}{if type \code{trips per day}, the parameter sets the number of columns to show on the facet plot.}

\item{palette}{if type \code{trips per day}, sets the palette to color edges}
}
\value{
For object of class \code{ftw_atmp} plots according to the type.
        For object of class \code{ftw_tmp} returns a network plot to compare between 2015 (left) and 2016 (right)
}
\description{
Explore the clipped Fietstelweek data using the library tmap with its \code{plot} mode.
}
