% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinces}
\alias{provinces}
\title{Provinces names in the Netherlands}
\format{A matrix with 1 columns and 12 rows containing the names of the
        Dutch provinces in alphabetical order.}
\source{
\url{http://www.imergis.nl/asp/47.asp}
}
\usage{
provinces
}
\description{
A matrix containing the provinces of
the Netherlands sorted in alfabetical order for referencial
use when clipping the Fietstelweek dataset.
}
\keyword{datasets}
