% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ftw}
\alias{read_ftw}
\alias{read_ftw}
\alias{print.ftw_atmp}
\alias{read_ftw}
\alias{print.ftw_tmp}
\alias{read_ftw}
\alias{summary.ftw_atmp}
\alias{read_ftw}
\alias{summary.ftw_tmp}
\title{Read clipped Fietstelweek data resulting of \code{clip_ftw}}
\usage{
read_ftw(directory, year, study_area)

\method{print}{ftw_atmp}(x, ...)

\method{print}{ftw_tmp}(x, ...)

\method{summary}{ftw_atmp}(object, ...)

\method{summary}{ftw_tmp}(object, ...)
}
\arguments{
\item{directory}{filepath to the directory where the Fietstelweek data is stored.
Same filepath as the one used for the \link[ftw]{clip_ftw} function.}

\item{year}{desired year of data, can either be \code{2015} or \code{2016}, which will return
an object of class \code{ftw_atmp} for "atemporal" analysis; or both \code{c(2015,2016)},
which will return an object of class \code{ftw_tmp} for "temporal" analysis
(comparison between years).}

\item{study_area}{name of the municipality or province the data was clipped to. The parameter is case sensitive
and prone to fail with spelling mistakes.}

\item{x}{object of class \code{ftw_atmp} or \code{ftw_tmp}}

\item{...}{ignored}

\item{object}{object of class \code{ftw_atmp} or \code{ftw_tmp}}
}
\value{
Returns either and object of class \code{ftw_atmp} for "atemporal" analysis when year is
        \code{2015} or \code{2016}, or into an object of class \code{ftw_tmp} for "temporal" analysis
        (comparison between years) when year is \code{c(2015,2016)}.

The summay method returns the total number of edges, nodes and trips of the data by year.
}
\description{
Reads the resulting files created by the \link[ftw]{clip_ftw} function and
loads them into an object of class \code{ftw_atmp} when loading data only for
one year, or into an object of class \code{ftw_tmp} when loading data for both years.
}
